Refactor: Organize files into folders and modularize DB functions
