+++ src/utilities/HttpError.ts
@@ -1,0 +1,11 @@
+/** @format */
+
+export class HttpError extends Error {
+        status: number;
+
+        constructor(status: number, message: string) {
+                super(message);
+                this.status = status;
+                Object.setPrototypeOf(this, HttpError.prototype);
+        }
+}
